requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0

sqlalchemy==2.0.36
aiosqlite==0.20.0

httpx==0.27.2
feedparser==6.0.11

trafilatura==1.12.2
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.3.0
markdownify==0.13.1

apscheduler==3.10.4
python-dotenv==1.0.1
